third_party/blink/web_tests/wpt_internal/overscroll-gestures/parsing/overscroll-position-valid.html

<!DOCTYPE html>
<meta charset="utf-8">

<title>Overscroll Position Parsing Test - valid</title>
<meta name="assert" content="overscroll-position supports the full grammar 'none | <dashed-ident>'.">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_valid_value("overscroll-position", "none", "none");
test_valid_value("overscroll-position", "nOnE", "none");
test_valid_value("overscroll-position", "--foo", "--foo");
</script>
